feat(cascade-api): change projects endpoints based on user system (#28)
* feat(cascade-api): change antomy of projects endpoint

* feat(migrations): add user_projects join table

* feat(migrations): add before update trigger for user_projects table

* feat(cascade-api): user_projects table access layer

* feat(cascade-api): user_projects table access layer

* refactor(cascade-api): rename table modules

* fix(cascade-api): set user_projects entity name

* feat(cascade-api): introduce helper Query module

* feat(cascade-api): implement handlers

* chore: run Headroom

* fix(cascade-api): delete unused module

* test(cascade-api): fix TestClient module

* test(cascade-api): authentication aware projects endpoint tests

* refactor(cascade-api): move Query modules under Database

* refactor(cascade-api): move Database effects to Depository

* Revert "refactor(cascade-api): move Database effects to Depository"

This reverts commit 423382c904241f2c677beb66e348a34375df0008.

* feat(cascade-api): authenticate /projects endpoints

* test(cascade-api): fix test client

* test(cascade-api): define semigroup and monoid instances for Validity

* test(cascade-api): add login client function

* test(cascade-api): add missing tests

* feat(cascade-api): assign user to project during project creation

* style: run Brittany

* fix(cascade-api): interpret database transaction correctly

* test(cascade-api): add pre-condition in create valid task for existing project

* feat(cascade-prelude): export `apply` from flow

* test(cascade-api): delete related tasks when delete project

* test(cascde-api): reset Postgres database by snapshot

* test(cascade-api): catch exception

* test(cascade-api): regulate sequential state machine tests

* refactor(cascade-api): fix warnings

* fix(cascade-api): remove forbidden response from projects endpoint

* style: remove redundant parens

* refactor(cascade-api): pipe and compositions!

* test(cascade-api): test incorrect and invalid login

* test(cascade-api): use Gen.replicateAtLeastOne instead of discard

* chore(Headroom): run

* test(cascade-api): produce valid user creatable

* refactor(cascade-api): remove redundant constraints

* style: fix GADTs alignment

+++ cascade-api/src/Cascade/Api/Network/Anatomy/Api/Projects.hs
@@ -10,14 +10,7 @@ Portability : POSIX
 !!! INSERT MODULE LONG DESCRIPTION !!!
 -}
 
-module Cascade.Api.Network.Anatomy.Api.Projects
-  ( Routes(..)
-  , CreateResponse
-  , GetAllResponse
-  , GetByIdResponse
-  , UpdateByIdResponse
-  , DeleteByIdResponse
-  ) where
+module Cascade.Api.Network.Anatomy.Api.Projects (Routes(..), GetByIdResponse, UpdateByIdResponse, DeleteByIdResponse) where
 
 import qualified Cascade.Api.Data.Project           as Project
 import qualified Cascade.Api.Network.Anatomy.Api.Projects.Tasks
@@ -26,22 +19,16 @@ import           Cascade.Api.Network.Anatomy.Prelude
 import qualified Cascade.Api.Servant.Response       as Response
 import           Data.Generics.Labels                ( )
 
-type CreateResponse = '[Response.Created Project.Readable]
+type GetByIdResponse = '[Response.Unauthorized , Response.NotFound , Response.Ok Project.Readable]
 
-type GetAllResponse = '[Response.Ok [Project.Readable]]
+type UpdateByIdResponse = '[Response.Unauthorized , Response.NotFound , Response.Ok Project.Readable]
 
-type GetByIdResponse = '[Response.Ok Project.Readable , Response.NotFound]
-
-type UpdateByIdResponse = '[Response.Ok Project.Readable , Response.NotFound]
-
-type DeleteByIdResponse = '[Response.Ok Project.Readable , Response.NotFound]
+type DeleteByIdResponse = '[Response.Unauthorized , Response.NotFound , Response.Ok Project.Readable]
 
 data Routes route = Routes
-  { create     :: route :- ReqBody '[JSON] Project.Creatable :> Post '[JSON] CreateResponse
-  , getAll     :: route :- Get '[JSON] GetAllResponse
-  , getById    :: route :- Capture "id" Project.Id :> Get '[JSON] GetByIdResponse
-  , updateById :: route :- Capture "id" Project.Id :> ReqBody '[JSON] Project.Updatable :> Patch '[JSON] UpdateByIdResponse
-  , deleteById :: route :- Capture "id" Project.Id :> Delete '[JSON] DeleteByIdResponse
+  { deleteById :: route :- Capture "id" Project.Id :> Auth :> Delete '[JSON] DeleteByIdResponse
+  , getById    :: route :- Capture "id" Project.Id :> Auth :> Get '[JSON] GetByIdResponse
   , tasks      :: route :- Capture "id" Project.Id :> "tasks" :> ToServantApi Api.Projects.Tasks.Routes
+  , updateById :: route :- Capture "id" Project.Id :> ReqBody '[JSON] Project.Updatable :> Auth :> Patch '[JSON] UpdateByIdResponse
   }
   deriving stock Generic
